# api/openapi.yaml
openapi: 3.0.3
info:
  title: Serenity ERP API
  description: |
    HIPAA-compliant API for Serenity Care Partners ERP system.
    
    ## Security
    - All endpoints require authentication via Bearer token
    - Role-based access control (RBAC) with attribute-based scoping (ABAC)
    - Row-level security enforced at database level
    - PHI endpoints require additional HIPAA compliance headers
    
    ## Rate Limiting
    - 1000 requests per minute per user
    - 10000 requests per minute per organization
    
    ## Data Classification
    Responses include `X-Data-Classification` header:
    - `public`: No sensitive information
    - `internal`: Internal business data
    - `confidential`: Sensitive business data
    - `phi`: Protected Health Information
    
  version: 1.0.0
  contact:
    name: Serenity Care Partners IT
    email: it@serenitycarepartners.com
  license:
    name: Proprietary
servers:
  - url: https://api.serenityerp.com/v1
    description: Production server
  - url: https://api-staging.serenityerp.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "123e4567-e89b-12d3-a456-426614174000"
    Timestamp:
      type: string
      format: date-time
      example: "2024-01-15T10:30:00Z"
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        details:
          type: object
      required: [error, message]
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
        pages:
          type: integer
      required: [page, limit, total, pages]
      
    # User and Authentication
    User:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        organizationId:
          $ref: '#/components/schemas/UUID'
        email:
          type: string
          format: email
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        role:
          type: string
          enum: [founder, security_officer, compliance_officer, finance_director, billing_manager, rcm_analyst, scheduler, field_supervisor, hr_manager, credentialing_specialist, it_admin, support_agent, caregiver, client, family, payer_auditor, ai_service]
        phone:
          type: string
          maxLength: 20
        isActive:
          type: boolean
        lastLogin:
          $ref: '#/components/schemas/Timestamp'
        mfaEnabled:
          type: boolean
        hourlyRate:
          type: number
          format: decimal
          multipleOf: 0.01
        preferences:
          type: object
        createdAt:
          $ref: '#/components/schemas/Timestamp'
        updatedAt:
          $ref: '#/components/schemas/Timestamp'
      required: [id, organizationId, email, firstName, lastName, role, isActive, mfaEnabled, createdAt, updatedAt]
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        role:
          type: string
          enum: [caregiver, scheduler, billing_manager, hr_manager, field_supervisor]
        phone:
          type: string
          maxLength: 20
        hourlyRate:
          type: number
          format: decimal
          multipleOf: 0.01
        hireDate:
          type: string
          format: date
      required: [email, firstName, lastName, role]
      
    # Client schemas (PHI)
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: role
          in: query
          schema:
            type: string
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of users
          headers:
            X-Data-Classification:
              schema:
                type: string
              example: internal
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
          
    post:
      tags: [Users]
      summary: Create new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          headers:
            X-Data-Classification:
              schema:
                type: string
              example: internal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...